refactor: Cloudify 추천 시스템 패키지 구조 정리 (parser, youtube 계층 분리)
