+++ api-spring/src/main/java/org/example/apispring/reco/service/parser/ConstraintParserService.java
@@ -1,4 +1,4 @@
-package org.example.apispring.reco.service;
+package org.example.apispring.reco.service.parser;
 
 import org.example.apispring.reco.dto.CanonicalTagQuery;
