+++ api-spring/src/main/java/org/example/apispring/reco/web/RecommendationController.java
@@ -3,7 +3,7 @@ package org.example.apispring.reco.web;
 import org.example.apispring.reco.dto.CanonicalTagQuery;
 import org.example.apispring.reco.dto.SongResponse;
 import org.example.apispring.reco.service.RecommendationService;
-import org.example.apispring.youtube.web.YouTubeService;
+import org.example.apispring.reco.service.youtube.YouTubeService;
 import org.example.apispring.youtube.web.YouTubeIdExtractor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
